requirements.txt
flask==3.0.0
gspread==6.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pyyaml==6.0.1
python-dateutil==2.8.2
